More gameboy stuff

+++ registers.h
@@ -3,20 +3,20 @@
 
 struct Registers
 {
-    long A;
-    long B;
-    long C;
-    long D;
-    long E;
-    long F;
-    long H;
-    long L;
-    long AF;
-    long SP;
-    long PC;
-    long BC;
-    long DE;
-    long HL;
+    uint8_t A;
+    uint8_t B;
+    uint8_t C;
+    uint8_t D;
+    uint8_t E;
+    uint8_t F;
+    uint8_t H;
+    uint8_t L;
+    uint16_t AF;
+    uint16_t SP;
+    uint16_t PC;
+    uint16_t BC;
+    uint16_t DE;
+    uint16_t HL;
 };
 
 #endif
